LayoutTests/media/modern-media-controls/placard-support/ipad/placard-support-pip-expected.txt

Testing the PlacardSupport behavior when entering picture-in-picture.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaController.controls.showsPlacard is true
PASS mediaController.controls.children.includes(mediaController.controls.pipPlacard) is true
PASS mediaController.controls.showsPlacard is false
PASS mediaController.controls.children.includes(mediaController.controls.pipPlacard) is false
PASS successfullyParsed is true

TEST COMPLETE
